pyproject.toml
[tool.poetry]
name = "agent-based"
version = "0.1.0"
description = ""
authors = ["Krzysztof P <krzypluc@gmail.com>"]
readme = "README.md"
packages = [{include = "agent_based"}]

[tool.poetry.dependencies]
python = "^3.9"
mesa = "^1.1.1"
pyyaml = "^6.0"
python-box = "^7.0.1"
openpyxl = "^3.1.2"
unidecode = "^1.3.6"
lxml = "^4.9.2"
geopandas = "^0.12.2"
pyarrow = "^11.0.0"
pyproj = "^3.4.1"
matplotlib = "^3.7.1"
pvlib = "^0.9.5"
pytz = "^2022.7.1"
python-dotenv = "^1.0.0"
joblib = "^1.2.0"
tables = "^3.8.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
